requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.2
click==7.1.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
python-dotenv==0.13.0
pytz==2019.3
redis==3.4.1
SQLAlchemy==1.3.16
vine==1.3.0
Werkzeug==1.0.1
zipp==3.1.0
